chore: move folders to src
