+++ src/_forms/equadiff.mdx
