+++ src/components/svg/HeroIllustration.jsx
