+++ src/components/svg/MathsIlllustration.jsx
