+++ src/styles/Navbar.module.css
@@ -49,3 +49,39 @@
  padding: 15px 35px;
  border-radius: 8px;
 }
+
+.nav__menu {
+ width: 60px;
+ height: 60px;
+ background: white;
+ border-radius: 50%;
+
+ pointer-events: none;
+ visibility: hidden;
+ display: none;
+}
+
+/* Responsive */
+
+@media screen and (max-width: 500px) {
+ .nav {
+  padding: 10px;
+ }
+
+ .nav__links ul {
+  display: inherit;
+  flex-direction: column;
+ }
+
+ .nav__featureBtn {
+  pointer-events: none;
+  visibility: hidden;
+  display: none;
+ }
+
+ .nav__menu {
+  pointer-events: all;
+  visibility: visible;
+  display: block;
+ }
+}
