+++ src/styles/globals.css
@@ -18,7 +18,7 @@ body {
 }
 
 :root {
- --blue: #6776ff;
+ --blue: #5162ff;
  --black: #0b053d;
 
  --maxWidth: 1260px;
